Splitted the cli test into two files.
